+++ src/steps/qualifiers.ts
@@ -16,11 +16,7 @@
 
 import { Node as PMNode, Slice } from 'prosemirror-model'
 import { Selection, Transaction } from 'prosemirror-state'
-import { ReplaceAroundStep, ReplaceStep, Step } from 'prosemirror-transform'
-
-import { TrackChangesAction } from '../actions'
-import { TrTrackingContext } from '../types/track'
-import { log } from '../utils/logger'
+import { ReplaceAroundStep, ReplaceStep } from 'prosemirror-transform'
 
 export const isDeleteStep = (step: ReplaceStep) =>
   step.from !== step.to && step.slice.content.size < step.to - step.from
@@ -50,7 +46,6 @@ export const isSplitStep = (step: ReplaceStep, selection: Selection, uiEvent: st
       (startOffset === parentSize && endOffset === parentSize)
     )
   }
-
   const {
     content: { firstChild, lastChild },
     openStart,
@@ -80,6 +75,12 @@ export const isWrapStep = (step: ReplaceAroundStep) =>
   step.slice.openStart === 0 &&
   step.slice.openEnd === 0
 
+/** Qualifies as a lift step when:
+    - there is a retained gap (captured original content that we insert)
+    - step.from < gapFrom  - meaning we remove content in front of the gap
+    - step.to > gapTo     - meaning we remove content after the gap
+    - nothing new is inserted: slice is empty
+  */
 export const isLiftStep = (step: ReplaceAroundStep) => {
   if (
     step.from < step.gapFrom &&
@@ -90,25 +91,19 @@ export const isLiftStep = (step: ReplaceAroundStep) => {
     return true
   }
   return false
-  /* qualifies as a lift step when:
-    - there is a retained gap (captured original content that we insert)
-    - step.from < gapFrom  - meaning we remove content in front of the gap
-    - step.to > gapTo     - meaning we remove content after the gap
-    - nothing new is inserted: slice is empty
-  */
 }
 
-export function isLiftStepForGap(
-  /*
+/**
     The step is a lift from an end of the step range.
     In other words it means that we removed a piece of content from the end of the step range,
     we then retained it and we put it at the start of the step range, e.g:
-      -> <p>
-      |  <ul>
-      |   <li>
-      ----- <p>
-              <p>
+    -> <p>
+    |  <ul>
+    |   <li>
+    ----- <p>
+            <p>
   */
+export function isLiftStepForGap(
   gap: {
     start: number
     end: number
@@ -119,31 +114,4 @@ export function isLiftStepForGap(
   to: number
 ) {
   return gap.start < gap.end && gap.insert === 0 && gap.end === to && !node.isText
-}
-
-export const isStructureSteps = (tr: Transaction) =>
-  tr.getMeta(TrackChangesAction.structuralChangeAction) &&
-  tr.steps.length === 2 &&
-  tr.steps[0] instanceof ReplaceStep &&
-  tr.steps[1] instanceof ReplaceStep
-
-export function processStepsBeforeTracking(
-  tr: Transaction,
-  trContext: TrTrackingContext,
-  processors: Array<(tr: Transaction, context: TrTrackingContext) => Step[] | void>
-) {
-  let steps: Step[] = []
-  processors.forEach((p) => {
-    const res = p(tr, trContext)
-    if (res) {
-      steps = res
-    }
-
-    if (steps.length < tr.steps.length) {
-      log.warn(
-        'Bug! A processor function filtered steps incorrectly. Filtered out steps should be replaced with null and not popped out of the array. Length and order has to be preserved'
-      )
-    }
-  })
-  return steps
 }
