+++ src/track-changes/track-transaction.ts
@@ -1,5 +1,5 @@
 /*!
- * © 2023 Atypon Systems LLC
+ * © 2025 Atypon Systems LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -30,24 +30,29 @@ import { diffChangeSteps } from '../change-steps/diffChangeSteps'
 import { processChangeSteps } from '../change-steps/processChangeSteps'
 import { updateChangeAttrs } from '../changes/updateChangeAttrs'
 import { getNodeTrackedData } from '../compute/nodeHelpers'
-import { CHANGE_STATUS } from '../types/change'
 import { ExposedReplaceStep } from '../types/pm'
-import { NewEmptyAttrs, TrTrackingContext } from '../types/track'
+import { TrTrackingContext } from '../types/track'
 import { log } from '../utils/logger'
-import { mapChangeSteps } from '../utils/mapChangeStep'
-import { excludeFromTracking, passThroughMeta, iterationIsValid } from '../utils/track-utils'
+import {
+  excludeFromTracking,
+  passThroughMeta,
+  iterationIsValid,
+  isStructuralChange,
+} from '../utils/track-utils'
 import { uuidv4 } from '../utils/uuidv4'
-import trackAttrsChange from './trackAttrsChange'
+import { NewEmptyAttrs } from '../attributes/types'
+import { createNewPendingAttrs } from '../attributes'
+import { trackReplaceStep } from '../steps/trackReplaceStep'
+import { isDeleteStep } from '../steps/qualifiers'
+import { trackReplaceAroundStep } from '../steps/trackReplaceAroundStep'
+import trackAttrsChange from '../steps/trackAttrsChange'
 import {
   trackAddMarkStep,
   trackAddNodeMarkStep,
   trackRemoveMarkStep,
   trackRemoveNodeMarkStep,
-} from './trackMarkSteps'
-import { trackReplaceAroundStep } from './trackReplaceAroundStep'
-import { trackReplaceStep } from './trackReplaceStep'
-import { fixAndSetSelectionAfterTracking } from './fixAndHandleSelection'
-import { isDeleteStep, isStructureSteps } from './utils'
+} from '../steps/trackMarkSteps'
+import { fixAndSetSelectionAfterTracking } from '../steps/fixAndHandleSelection'
 
 /**
  * Inverts transactions to wrap their contents/operations with track data instead
@@ -66,7 +71,6 @@ import { isDeleteStep, isStructureSteps } from './utils'
  * @param authorID User id
  * @returns newTr that inverts the initial tr and applies track attributes/marks
  */
-
 export function trackTransaction(
   tr: Transaction,
   oldState: EditorState,
@@ -75,25 +79,7 @@ export function trackTransaction(
   clearedSteps: Step[],
   trContext: TrTrackingContext
 ) {
-  /*
-      1. Import all change owners
-      2. For each step, call a change owner method that to corresponds to that step
-      i.e.
-      else if (step instanceof ReplaceStep) {
-        owners.forEach(o => o.onReplaceStep?(step, tr)          
-      }
-    */
-
-  const emptyAttrs: NewEmptyAttrs = {
-    authorID,
-    reviewedByID: null,
-    createdAt: tr.time,
-    updatedAt: tr.time,
-    statusUpdateAt: 0, // has to be zero as first so changes are not differeniated at start
-    status: CHANGE_STATUS.pending,
-  }
-
-  // Check for indentation metadata and treat it like a move operation
+  const emptyAttrs: NewEmptyAttrs = createNewPendingAttrs(tr.time, authorID)
 
   // mapping for deleted content, that was inserted before
   const deletedNodeMapping = new Mapping()
@@ -126,7 +112,7 @@ export function trackTransaction(
       step adds content before (in terms of position in the doc) the first step, the plugin will attempt to insert tracked replacement for the first change at a position
       that corresponds to the first change position if the second change (second in time but occuring earlier in doc) never occured.
       */
-      if (isDeleteStep(step) || isStructureSteps(tr)) {
+      if (isDeleteStep(step) || isStructuralChange(tr)) {
         thisStepMapping = deletedNodeMapping
       }
       let [steps] = trackReplaceStep(i, oldState, newTr, emptyAttrs, tr, thisStepMapping, trContext)
@@ -138,7 +124,7 @@ export function trackTransaction(
           continue
         }
       }
-      log.info('TRACK REPLACE CHANGES: ', steps)
+      log.info('TRACK REPLACE CHANGES: ', [...steps])
       steps = diffChangeSteps(steps)
       log.info('DIFFED STEPS: ', steps)
 
@@ -181,9 +167,6 @@ export function trackTransaction(
     } else if (step instanceof AddNodeMarkStep) {
       trackAddNodeMarkStep(step, emptyAttrs, newTr, tr.docs[i])
     }
-    // TODO: here we could check whether adjacent inserts & deletes cancel each other out.
-    // However, this should not be done by diffing and only matching node or char by char instead since
-    // it's A easier and B more intuitive to user.
   }
   newTr = passThroughMeta(tr, newTr)
   newTr = fixAndSetSelectionAfterTracking(newTr, tr, deletedNodeMapping, trContext)
